% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{as.data.frame.asa_audit_result}
\alias{as.data.frame.asa_audit_result}
\title{Convert asa_audit_result to Data Frame}
\usage{
\method{as.data.frame}{asa_audit_result}(x, ...)
}
\arguments{
\item{x}{An asa_audit_result object}

\item{...}{Additional arguments (ignored)}
}
\value{
The audited data.frame with audit columns
}
\description{
Convert asa_audit_result to Data Frame
}
