% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{as.data.frame.asa_enumerate_result}
\alias{as.data.frame.asa_enumerate_result}
\title{Convert asa_enumerate_result to Data Frame}
\usage{
\method{as.data.frame}{asa_enumerate_result}(x, ...)
}
\arguments{
\item{x}{An asa_enumerate_result object}

\item{...}{Additional arguments (ignored)}
}
\value{
The data data.frame from the result
}
\description{
Convert asa_enumerate_result to Data Frame
}
