% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public__s3_methods.R
\name{as.data.frame.asa_evaluation_result}
\alias{as.data.frame.asa_evaluation_result}
\title{Convert asa_evaluation_result to Data Frame}
\usage{
\method{as.data.frame}{asa_evaluation_result}(x, ...)
}
\arguments{
\item{x}{An asa_evaluation_result object}

\item{...}{Additional arguments (ignored)}
}
\value{
The per-task evaluation records as a data.frame
}
\description{
Convert asa_evaluation_result to Data Frame
}
