% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{as.data.frame.asa_result}
\alias{as.data.frame.asa_result}
\title{Convert asa_result to Data Frame}
\usage{
\method{as.data.frame}{asa_result}(x, ...)
}
\arguments{
\item{x}{An asa_result object}

\item{...}{Additional arguments (ignored)}
}
\value{
A single-row data frame
}
\description{
Convert asa_result to Data Frame
}
