% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_ADAPTIVE_RATE_DECREASE}
\alias{ASA_ADAPTIVE_RATE_DECREASE}
\title{Adaptive Rate Decrease Factor (on success streak)}
\usage{
ASA_ADAPTIVE_RATE_DECREASE
}
\description{
Multiply delays by this factor after 10 consecutive successes.
}
\keyword{internal}
