% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_ADAPTIVE_RATE_MAX}
\alias{ASA_ADAPTIVE_RATE_MAX}
\title{Adaptive Rate Maximum Multiplier}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ASA_ADAPTIVE_RATE_MAX
}
\description{
Cap on delay multiplier to prevent excessive slowdown.
}
\keyword{internal}
