% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_ADAPTIVE_RATE_MIN}
\alias{ASA_ADAPTIVE_RATE_MIN}
\title{Adaptive Rate Minimum Multiplier}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ASA_ADAPTIVE_RATE_MIN
}
\description{
Floor on delay multiplier to maintain some speed.
}
\keyword{internal}
