% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_CIRCUIT_BREAKER_COOLDOWN}
\alias{ASA_CIRCUIT_BREAKER_COOLDOWN}
\title{Circuit Breaker Cooldown Period (seconds to wait when tripped)}
\usage{
ASA_CIRCUIT_BREAKER_COOLDOWN
}
\description{
Circuit Breaker Cooldown Period (seconds to wait when tripped)
}
\keyword{internal}
