% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_CIRCUIT_BREAKER_ENABLED}
\alias{ASA_CIRCUIT_BREAKER_ENABLED}
\title{Enable Circuit Breaker by default}
\usage{
ASA_CIRCUIT_BREAKER_ENABLED
}
\description{
Enable Circuit Breaker by default
}
\keyword{internal}
