% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\docType{data}
\name{ASA_DEFAULT_AGENT_BACKEND}
\alias{ASA_DEFAULT_AGENT_BACKEND}
\title{Default Agent Backend}
\format{
An object of class \code{character} of length 1.
}
\usage{
ASA_DEFAULT_AGENT_BACKEND
}
\description{
Default Agent Backend
}
\keyword{internal}
