% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_DEFAULT_BEDROCK_MODEL}
\alias{ASA_DEFAULT_BEDROCK_MODEL}
\title{Default Bedrock Model}
\format{
An object of class \code{character} of length 1.
}
\usage{
ASA_DEFAULT_BEDROCK_MODEL
}
\description{
Default Bedrock Model
}
\keyword{internal}
