% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_DEFAULT_BUDGET_TOKENS}
\alias{ASA_DEFAULT_BUDGET_TOKENS}
\title{Default Budget: Tokens}
\format{
An object of class \code{integer} of length 1.
}
\usage{
ASA_DEFAULT_BUDGET_TOKENS
}
\description{
Default Budget: Tokens
}
\keyword{internal}
