% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_DEFAULT_CAPTCHA_BACKOFF_BASE}
\alias{ASA_DEFAULT_CAPTCHA_BACKOFF_BASE}
\title{Default CAPTCHA Backoff Base Multiplier}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ASA_DEFAULT_CAPTCHA_BACKOFF_BASE
}
\description{
Aggressive backoff on CAPTCHA: 5.0x multiplier.
  Results in 5s, 10s, 15s delays on successive CAPTCHA encounters.
}
\keyword{internal}
