% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_DEFAULT_CONDA_ENV}
\alias{ASA_DEFAULT_CONDA_ENV}
\title{Default Conda Environment}
\format{
An object of class \code{character} of length 1.
}
\usage{
ASA_DEFAULT_CONDA_ENV
}
\description{
Default Conda Environment
}
\keyword{internal}
