% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_DEFAULT_GEMINI_MODEL}
\alias{ASA_DEFAULT_GEMINI_MODEL}
\title{Default Gemini Model}
\format{
An object of class \code{character} of length 1.
}
\usage{
ASA_DEFAULT_GEMINI_MODEL
}
\description{
Default Gemini Model
}
\keyword{internal}
