% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_DEFAULT_INVOKE_RETRY_BACKOFF}
\alias{ASA_DEFAULT_INVOKE_RETRY_BACKOFF}
\title{Default Agent Invoke Retry Backoff Multiplier}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ASA_DEFAULT_INVOKE_RETRY_BACKOFF
}
\description{
Default Agent Invoke Retry Backoff Multiplier
}
\keyword{internal}
