% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_DEFAULT_INVOKE_RETRY_DELAY}
\alias{ASA_DEFAULT_INVOKE_RETRY_DELAY}
\title{Default Agent Invoke Retry Delay (seconds)}
\usage{
ASA_DEFAULT_INVOKE_RETRY_DELAY
}
\description{
Default Agent Invoke Retry Delay (seconds)
}
\keyword{internal}
