% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_DEFAULT_INVOKE_RETRY_JITTER}
\alias{ASA_DEFAULT_INVOKE_RETRY_JITTER}
\title{Default Agent Invoke Retry Jitter Factor}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ASA_DEFAULT_INVOKE_RETRY_JITTER
}
\description{
Fractional jitter applied to retry delays (0.25 = +/-25%).
}
\keyword{internal}
