% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_DEFAULT_MAX_ROUNDS}
\alias{ASA_DEFAULT_MAX_ROUNDS}
\title{Default Max Rounds for Enumeration}
\usage{
ASA_DEFAULT_MAX_ROUNDS
}
\description{
Default Max Rounds for Enumeration
}
\keyword{internal}
