% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_DEFAULT_MEMORY_KEEP_RECENT}
\alias{ASA_DEFAULT_MEMORY_KEEP_RECENT}
\title{Default Exchanges to Keep After Folding}
\usage{
ASA_DEFAULT_MEMORY_KEEP_RECENT
}
\description{
Default Exchanges to Keep After Folding
}
\keyword{internal}
