% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_DEFAULT_MEMORY_SUMMARIZER_MAX_OUTPUT_TOKENS}
\alias{ASA_DEFAULT_MEMORY_SUMMARIZER_MAX_OUTPUT_TOKENS}
\title{Default Memory Summarizer Output Cap (tokens)}
\usage{
ASA_DEFAULT_MEMORY_SUMMARIZER_MAX_OUTPUT_TOKENS
}
\description{
Default Memory Summarizer Output Cap (tokens)
}
\keyword{internal}
