% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\docType{data}
\name{ASA_DEFAULT_MIN_FOLD_BATCH}
\alias{ASA_DEFAULT_MIN_FOLD_BATCH}
\title{Default Minimum Fold Batch Size}
\format{
An object of class \code{integer} of length 1.
}
\usage{
ASA_DEFAULT_MIN_FOLD_BATCH
}
\description{
Minimum number of messages that must be eligible for folding
  before a fold is triggered. Prevents tiny folds (2-3 messages) that cost
  ~7K tokens in summarizer overhead but compress almost nothing.
}
\keyword{internal}
