% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_DEFAULT_NOVELTY_WINDOW}
\alias{ASA_DEFAULT_NOVELTY_WINDOW}
\title{Default Novelty Window}
\format{
An object of class \code{integer} of length 1.
}
\usage{
ASA_DEFAULT_NOVELTY_WINDOW
}
\description{
Default Novelty Window
}
\keyword{internal}
