% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\docType{data}
\name{ASA_DEFAULT_OLLAMA_MODEL}
\alias{ASA_DEFAULT_OLLAMA_MODEL}
\title{Default Ollama Model}
\format{
An object of class \code{character} of length 1.
}
\usage{
ASA_DEFAULT_OLLAMA_MODEL
}
\description{
Default Ollama Model
}
\keyword{internal}
