% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_DEFAULT_OM_ASYNC_PREBUFFER}
\alias{ASA_DEFAULT_OM_ASYNC_PREBUFFER}
\title{Default OM Async Prebuffering}
\usage{
ASA_DEFAULT_OM_ASYNC_PREBUFFER
}
\description{
Default OM Async Prebuffering
}
\keyword{internal}
