% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_DEFAULT_OM_BLOCK_AFTER}
\alias{ASA_DEFAULT_OM_BLOCK_AFTER}
\title{Default OM Block-After Ratio}
\usage{
ASA_DEFAULT_OM_BLOCK_AFTER
}
\description{
Default OM Block-After Ratio
}
\keyword{internal}
