% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_DEFAULT_OM_BUFFER_ACTIVATION}
\alias{ASA_DEFAULT_OM_BUFFER_ACTIVATION}
\title{Default OM Buffer Activation Ratio}
\usage{
ASA_DEFAULT_OM_BUFFER_ACTIVATION
}
\description{
Default OM Buffer Activation Ratio
}
\keyword{internal}
