% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_DEFAULT_OM_OBSERVATION_TOKENS}
\alias{ASA_DEFAULT_OM_OBSERVATION_TOKENS}
\title{Default OM Observation Trigger Budget (approx tokens)}
\usage{
ASA_DEFAULT_OM_OBSERVATION_TOKENS
}
\description{
Default OM Observation Trigger Budget (approx tokens)
}
\keyword{internal}
