% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_DEFAULT_OM_REFLECTION_TOKENS}
\alias{ASA_DEFAULT_OM_REFLECTION_TOKENS}
\title{Default OM Reflection Trigger Budget (approx tokens)}
\usage{
ASA_DEFAULT_OM_REFLECTION_TOKENS
}
\description{
Default OM Reflection Trigger Budget (approx tokens)
}
\keyword{internal}
