% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_DEFAULT_PLATEAU_ROUNDS}
\alias{ASA_DEFAULT_PLATEAU_ROUNDS}
\title{Default Plateau Rounds for Stopping}
\usage{
ASA_DEFAULT_PLATEAU_ROUNDS
}
\description{
Default Plateau Rounds for Stopping
}
\keyword{internal}
