% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_DEFAULT_TEMPERATURES}
\alias{ASA_DEFAULT_TEMPERATURES}
\title{Default Temperatures by Backend}
\usage{
ASA_DEFAULT_TEMPERATURES
}
\description{
Default Temperatures by Backend
}
\keyword{internal}
