% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_DEFAULT_USE_OBSERVATIONAL_MEMORY}
\alias{ASA_DEFAULT_USE_OBSERVATIONAL_MEMORY}
\title{Default Observational Memory Enabled}
\usage{
ASA_DEFAULT_USE_OBSERVATIONAL_MEMORY
}
\description{
Default Observational Memory Enabled
}
\keyword{internal}
