% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_HUMANIZE_TIMING}
\alias{ASA_HUMANIZE_TIMING}
\title{Enable Humanized Timing (random jitter on delays)}
\usage{
ASA_HUMANIZE_TIMING
}
\description{
Enable Humanized Timing (random jitter on delays)
}
\keyword{internal}
