% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_JITTER_FACTOR}
\alias{ASA_JITTER_FACTOR}
\title{Jitter Factor for random timing variation}
\usage{
ASA_JITTER_FACTOR
}
\description{
Jitter Factor for random timing variation
}
\keyword{internal}
