% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_OPENCODE_DEFAULT_AGENT}
\alias{ASA_OPENCODE_DEFAULT_AGENT}
\title{Default OpenCode Agent Name}
\usage{
ASA_OPENCODE_DEFAULT_AGENT
}
\description{
Default OpenCode Agent Name
}
\keyword{internal}
