% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_OPENCODE_OUTER_MODEL}
\alias{ASA_OPENCODE_OUTER_MODEL}
\title{Fixed Outer Model for OpenCode Anthropic-Compatible Gateway}
\usage{
ASA_OPENCODE_OUTER_MODEL
}
\description{
Fixed Outer Model for OpenCode Anthropic-Compatible Gateway
}
\keyword{internal}
