% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_PROACTIVE_ROTATION_ENABLED}
\alias{ASA_PROACTIVE_ROTATION_ENABLED}
\title{Enable Proactive Tor Circuit Rotation}
\usage{
ASA_PROACTIVE_ROTATION_ENABLED
}
\description{
When TRUE, rotate Tor circuit every N requests (not just on error).
}
\keyword{internal}
