% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_PROACTIVE_ROTATION_INTERVAL}
\alias{ASA_PROACTIVE_ROTATION_INTERVAL}
\title{Proactive Rotation Interval (requests)}
\usage{
ASA_PROACTIVE_ROTATION_INTERVAL
}
\description{
Rotate Tor circuit every 15 requests to get fresh exit node IP.
}
\keyword{internal}
