% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_RATE_LIMIT_PROACTIVE}
\alias{ASA_RATE_LIMIT_PROACTIVE}
\title{Enable Proactive Rate Limiting (default: TRUE)}
\usage{
ASA_RATE_LIMIT_PROACTIVE
}
\description{
Enable Proactive Rate Limiting (default: TRUE)
}
\keyword{internal}
