% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_RECURSION_LIMIT_FOLDING}
\alias{ASA_RECURSION_LIMIT_FOLDING}
\title{Recursion Limit with Memory Folding}
\usage{
ASA_RECURSION_LIMIT_FOLDING
}
\description{
Recursion Limit with Memory Folding
}
\keyword{internal}
