% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_RECURSION_LIMIT_STANDARD}
\alias{ASA_RECURSION_LIMIT_STANDARD}
\title{Recursion Limit without Memory Folding}
\format{
An object of class \code{integer} of length 1.
}
\usage{
ASA_RECURSION_LIMIT_STANDARD
}
\description{
Recursion Limit without Memory Folding
}
\keyword{internal}
