% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_SESSION_RESET_ENABLED}
\alias{ASA_SESSION_RESET_ENABLED}
\title{Enable Session Reset}
\format{
An object of class \code{logical} of length 1.
}
\usage{
ASA_SESSION_RESET_ENABLED
}
\description{
When TRUE, periodically reset session identity to avoid fingerprinting.
}
\keyword{internal}
