% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_SUPPORTED_BACKENDS}
\alias{ASA_SUPPORTED_BACKENDS}
\title{Supported Backends}
\format{
An object of class \code{character} of length 5.
}
\usage{
ASA_SUPPORTED_BACKENDS
}
\description{
Supported Backends
}
\keyword{internal}
