% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_TOR_BAD_TTL}
\alias{ASA_TOR_BAD_TTL}
\title{Tor Exit Bad TTL (seconds)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ASA_TOR_BAD_TTL
}
\description{
How long to keep a bad/tainted exit before allowing reuse.
}
\keyword{internal}
