% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_TOR_IP_CACHE_TTL}
\alias{ASA_TOR_IP_CACHE_TTL}
\title{Tor Exit IP Cache TTL (seconds)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ASA_TOR_IP_CACHE_TTL
}
\description{
How long to cache exit IP lookups before refreshing.
}
\keyword{internal}
