% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_TOR_MAX_ROTATION_ATTEMPTS}
\alias{ASA_TOR_MAX_ROTATION_ATTEMPTS}
\title{Tor Rotation Attempts when Exit is Bad/Overused}
\usage{
ASA_TOR_MAX_ROTATION_ATTEMPTS
}
\description{
Tor Rotation Attempts when Exit is Bad/Overused
}
\keyword{internal}
