% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_TOR_OVERUSE_DECAY}
\alias{ASA_TOR_OVERUSE_DECAY}
\title{Tor Exit Overuse Decay Window (seconds)}
\usage{
ASA_TOR_OVERUSE_DECAY
}
\description{
Time window for counting recent uses before decaying counts.
}
\keyword{internal}
