% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_TOR_OVERUSE_THRESHOLD}
\alias{ASA_TOR_OVERUSE_THRESHOLD}
\title{Tor Exit Overuse Threshold}
\usage{
ASA_TOR_OVERUSE_THRESHOLD
}
\description{
Maximum recent uses before a good exit is considered overloaded.
}
\keyword{internal}
