% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow__benchmarks.R
\name{benchmark_from_yaml}
\alias{benchmark_from_yaml}
\title{Load an ASA Benchmark from YAML}
\usage{
benchmark_from_yaml(path)
}
\arguments{
\item{path}{Path to YAML benchmark file}
}
\value{
An object of class \code{asa_benchmark}
}
\description{
Reads a benchmark definition from a YAML file.
}
