% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_task.R
\name{build_prompt}
\alias{build_prompt}
\title{Build a Task Prompt from Template}
\usage{
build_prompt(template, ...)
}
\arguments{
\item{template}{A character string with placeholders in the form \code{{variable_name}}}

\item{...}{Named arguments to substitute into the template}
}
\value{
A formatted prompt string
}
\description{
Creates a formatted prompt by substituting variables into a template.
}
\examples{
\dontrun{
prompt <- build_prompt(
  template = "Find information about {{name}} in {{country}} during {{year}}",
  name = "Marie Curie",
  country = "France",
  year = 1903
)
}

}
