% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clean_whitespace}
\alias{clean_whitespace}
\title{Clean Whitespace}
\usage{
clean_whitespace(x)
}
\arguments{
\item{x}{Character string}
}
\value{
Cleaned string
}
\description{
Normalizes whitespace in a string by collapsing multiple spaces and
trimming leading/trailing whitespace.
}
\keyword{internal}
