% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{decode_html}
\alias{decode_html}
\title{Decode HTML Entities}
\usage{
decode_html(x)
}
\arguments{
\item{x}{Character string with HTML entities}
}
\value{
Decoded string
}
\description{
Converts HTML entities to their character equivalents.
}
\keyword{internal}
