% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.agent_matches_config}
\alias{.agent_matches_config}
\title{Check Whether an Agent Matches an asa_config}
\usage{
.agent_matches_config(agent, config)
}
\description{
Check Whether an Agent Matches an asa_config
}
\keyword{internal}
