% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess__process_outputs.R
\name{.aggregate_langgraph_timings}
\alias{.aggregate_langgraph_timings}
\title{Aggregate LangGraph timings by node for compact reporting}
\usage{
.aggregate_langgraph_timings(langgraph_timings = list())
}
\description{
Aggregate LangGraph timings by node for compact reporting
}
\keyword{internal}
