% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess__process_outputs.R
\name{.as_scalar_minutes}
\alias{.as_scalar_minutes}
\title{Coerce a value to scalar minutes (double)}
\usage{
.as_scalar_minutes(value, default = NA_real_)
}
\description{
Coerce a value to scalar minutes (double)
}
\keyword{internal}
