% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{.asa_option}
\alias{.asa_option}
\title{Get Package Option or Default}
\usage{
.asa_option(name, default)
}
\arguments{
\item{name}{Option name (without "asa." prefix)}

\item{default}{Default value if option not set}
}
\value{
Option value or default
}
\description{
Returns the value of an asa package option, or the default if not set.
Options can be set via options(asa.option_name = value).
}
\keyword{internal}
