% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent__initialize.R
\name{.azure_openai_base_url_from_env}
\alias{.azure_openai_base_url_from_env}
\title{Resolve Azure OpenAI Endpoint from Environment}
\usage{
.azure_openai_base_url_from_env()
}
\description{
Resolve Azure OpenAI Endpoint from Environment
}
\keyword{internal}
