% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra__runtime_config.R
\name{.compact_runtime_snapshot_args}
\alias{.compact_runtime_snapshot_args}
\title{Remove NULL Entries from Snapshot Arguments}
\usage{
.compact_runtime_snapshot_args(values)
}
\description{
Remove NULL Entries from Snapshot Arguments
}
\keyword{internal}
