% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow__run_task.R
\name{.deep_merge_named_lists}
\alias{.deep_merge_named_lists}
\title{Recursively Merge Named Lists}
\usage{
.deep_merge_named_lists(base, override)
}
\arguments{
\item{base}{Base list.}

\item{override}{Override list.}
}
\value{
A merged list.
}
\description{
Internal helper that recursively merges nested named lists.
}
\keyword{internal}
