% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent__free_code.R
\name{.free_code_expected_schema_to_json_schema}
\alias{.free_code_expected_schema_to_json_schema}
\title{Convert ASA expected_schema descriptors to JSON Schema for free-code}
\usage{
.free_code_expected_schema_to_json_schema(expected_schema)
}
\description{
Convert ASA expected_schema descriptors to JSON Schema for free-code
}
\keyword{internal}
