% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent__free_code.R
\name{.free_code_mcp_env}
\alias{.free_code_mcp_env}
\title{Build environment for the MCP search server}
\usage{
.free_code_mcp_env(
  config,
  python_path,
  allow_read_webpages = NULL,
  auto_openwebpage_policy = NULL
)
}
\description{
Build environment for the MCP search server
}
\keyword{internal}
