% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent__free_code.R
\name{.free_code_parse_output}
\alias{.free_code_parse_output}
\title{Parse free-code JSON stdout}
\usage{
.free_code_parse_output(stdout)
}
\description{
Parse free-code JSON stdout
}
\keyword{internal}
