% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent__free_code.R
\name{.free_code_terminate_process}
\alias{.free_code_terminate_process}
\title{Terminate a background process created for free-code integration}
\usage{
.free_code_terminate_process(proc)
}
\description{
Terminate a background process created for free-code integration
}
\keyword{internal}
