% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{.get_default_model_for_backend}
\alias{.get_default_model_for_backend}
\title{Get Default Model for Backend}
\usage{
.get_default_model_for_backend(backend)
}
\description{
Get Default Model for Backend
}
\keyword{internal}
