% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asa-package.R
\name{.get_python_path}
\alias{.get_python_path}
\title{Get Package Python Module Path}
\usage{
.get_python_path()
}
\value{
Character string with the path to inst/python
}
\description{
Returns the path to the Python modules shipped with the package.
}
\keyword{internal}
