% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra__heartbeat.R
\name{.heartbeat_process_alive}
\alias{.heartbeat_process_alive}
\title{Check if Heartbeat Process is Alive}
\usage{
.heartbeat_process_alive(pid)
}
\arguments{
\item{pid}{Process id.}
}
\value{
TRUE when process exists, otherwise FALSE.
}
\description{
Check if Heartbeat Process is Alive
}
\keyword{internal}
