% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra__heartbeat.R
\name{.heartbeat_set_phase}
\alias{.heartbeat_set_phase}
\title{Update Heartbeat Phase File}
\usage{
.heartbeat_set_phase(hb, phase = "running", detail = NA_character_)
}
\arguments{
\item{hb}{Heartbeat handle.}

\item{phase}{Phase token.}

\item{detail}{Additional detail token.}
}
\value{
Invisible NULL.
}
\description{
Update Heartbeat Phase File
}
\keyword{internal}
