% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra__heartbeat.R
\name{.heartbeat_wait_for_exit}
\alias{.heartbeat_wait_for_exit}
\title{Wait for Heartbeat Process Exit}
\usage{
.heartbeat_wait_for_exit(pid, timeout_sec = 2)
}
\arguments{
\item{pid}{Process id.}

\item{timeout_sec}{Max seconds to wait.}
}
\value{
TRUE when process exited before timeout, otherwise FALSE.
}
\description{
Wait for Heartbeat Process Exit
}
\keyword{internal}
