% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_outputs.R
\name{.make_json_string_wrapper_safe}
\alias{.make_json_string_wrapper_safe}
\title{Make a string safe to wrap as a JSON string literal}
\usage{
.make_json_string_wrapper_safe(x)
}
\arguments{
\item{x}{Character scalar}
}
\description{
Ensures every double quote is escaped (i.e., preceded by an odd number of
backslashes), so `jsonlite::fromJSON(paste0('"', x, '"'))` can be used to
decode one layer of trace-escaped content.
}
\keyword{internal}
