% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{.normalize_backend_model}
\alias{.normalize_backend_model}
\title{Normalize Backend/Model Shorthand}
\usage{
.normalize_backend_model(backend, model = NULL)
}
\description{
Normalize Backend/Model Shorthand
}
\keyword{internal}
