% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess__process_outputs.R
\name{.normalize_progress_status}
\alias{.normalize_progress_status}
\title{Summarize plan history for action timelines}
\usage{
.normalize_progress_status(x)
}
\description{
Summarize plan history for action timelines
}
\keyword{internal}
