% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow__research_enumeration.R
\name{.normalize_schema}
\alias{.normalize_schema}
\title{Normalize Schema Input}
\usage{
.normalize_schema(schema, query, verbose)
}
\description{
Normalize Schema Input
}
\keyword{internal}
