% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent__opencode.R
\name{.opencode_model_ref}
\alias{.opencode_model_ref}
\title{Convert an OpenCode model id into provider/model form}
\usage{
.opencode_model_ref(outer_model = ASA_OPENCODE_OUTER_MODEL)
}
\description{
Convert an OpenCode model id into provider/model form
}
\keyword{internal}
