% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent__opencode.R
\name{.opencode_parse_export_stdout}
\alias{.opencode_parse_export_stdout}
\title{Parse stdout from `opencode export`}
\usage{
.opencode_parse_export_stdout(stdout)
}
\description{
Parse stdout from `opencode export`
}
\keyword{internal}
