% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_outputs.R
\name{.parse_trace_json_messages}
\alias{.parse_trace_json_messages}
\title{Parse Structured Trace JSON (asa_trace_v1)}
\usage{
.parse_trace_json_messages(text)
}
\description{
Parse Structured Trace JSON (asa_trace_v1)
}
\keyword{internal}
