% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow__research_enumeration.R
\name{.process_research_results}
\alias{.process_research_results}
\title{Process Research Results}
\usage{
.process_research_results(result, schema_dict, include_provenance)
}
\description{
Process Research Results
}
\keyword{internal}
