% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow__research_enumeration.R
\name{.resume_research}
\alias{.resume_research}
\title{Resume Research from Checkpoint}
\usage{
.resume_research(checkpoint_file, verbose)
}
\description{
Resume Research from Checkpoint
}
\keyword{internal}
