% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_agent.R
\name{.run_agent}
\alias{.run_agent}
\title{Run the ASA Agent (Internal)}
\usage{
.run_agent(
  prompt,
  agent = NULL,
  recursion_limit = NULL,
  expected_schema = NULL,
  thread_id = NULL,
  field_status = NULL,
  budget_state = NULL,
  search_budget_limit = NULL,
  unknown_after_searches = NULL,
  finalize_on_all_fields_resolved = NULL,
  verbose = FALSE
)
}
\arguments{
\item{prompt}{The prompt to send to the agent}

\item{agent}{An asa_agent object}

\item{recursion_limit}{Maximum number of agent steps}

\item{verbose}{Print status messages}
}
\value{
An object of class \code{asa_response}
}
\description{
Internal function that invokes the search agent with a prompt.
Users should use \code{\link{run_task}} instead.
}
\keyword{internal}
