% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra__system_cmd.R
\name{.run_processx}
\alias{.run_processx}
\title{Run processx::run through a mockable internal seam}
\usage{
.run_processx(...)
}
\description{
Run processx::run through a mockable internal seam
}
\keyword{internal}
