% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow__research_enumeration.R
\name{.run_research}
\alias{.run_research}
\title{Run Research (Non-Streaming)}
\usage{
.run_research(graph, query, schema_dict, config_dict)
}
\description{
Run Research (Non-Streaming)
}
\keyword{internal}
