% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow__research_enumeration.R
\name{.save_checkpoint}
\alias{.save_checkpoint}
\title{Save Checkpoint}
\usage{
.save_checkpoint(
  result = NULL,
  query = NULL,
  schema_dict = NULL,
  config_dict = NULL,
  checkpoint_file,
  checkpoint_payload = NULL
)
}
\description{
Save Checkpoint
}
\keyword{internal}
