% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess__process_outputs.R
\name{.select_plan_candidate}
\alias{.select_plan_candidate}
\title{Choose the best plan candidate for summary}
\usage{
.select_plan_candidate(history_candidate = NULL, plan_candidate = NULL)
}
\description{
Choose the best plan candidate for summary
}
\keyword{internal}
