% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess__process_outputs.R
\name{.select_sources}
\alias{.select_sources}
\title{Select sources from an ordered vector (1-indexed)}
\usage{
.select_sources(result, source = NULL)
}
\description{
Select sources from an ordered vector (1-indexed)
}
\keyword{internal}
