% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess__process_outputs.R
\name{.summarize_field_status}
\alias{.summarize_field_status}
\title{Summarize field resolution status for investigators}
\usage{
.summarize_field_status(field_status = list(), max_chars = 88L)
}
\description{
Summarize field resolution status for investigators
}
\keyword{internal}
