% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess__process_outputs.R
\name{.summarize_plan_history}
\alias{.summarize_plan_history}
\title{Summarize plan history for action timelines}
\usage{
.summarize_plan_history(plan_history, plan = list(), max_chars = 88L)
}
\description{
Summarize plan history for action timelines
}
\keyword{internal}
