% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent__initialize.R
\name{.sync_provider_backend_env}
\alias{.sync_provider_backend_env}
\title{Synchronize Provider Marker Environment into R and Python}
\usage{
.sync_provider_backend_env(
  backend,
  azure_base_url = NULL,
  azure_deployment = NULL
)
}
\description{
Synchronize Provider Marker Environment into R and Python
}
\keyword{internal}
