% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_asa_config}
\alias{.validate_asa_config}
\title{Validate Optional asa_config Argument}
\usage{
.validate_asa_config(config, param_name = "config")
}
\arguments{
\item{config}{Value to check (NULL or asa_config)}

\item{param_name}{Name for error message}
}
\description{
Validate Optional asa_config Argument
}
\keyword{internal}
