% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_asa_response}
\alias{.validate_asa_response}
\title{Validate S3 Constructor: asa_response}
\usage{
.validate_asa_response(
  message,
  status_code,
  raw_response,
  trace,
  elapsed_time,
  prompt,
  fold_stats = list()
)
}
\description{
Validate S3 Constructor: asa_response
}
\keyword{internal}
