% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_asa_result}
\alias{.validate_asa_result}
\title{Validate S3 Constructor: asa_result}
\usage{
.validate_asa_result(prompt, message, parsed, raw_output, elapsed_time, status)
}
\description{
Validate S3 Constructor: asa_result
}
\keyword{internal}
