% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_build_prompt}
\alias{.validate_build_prompt}
\title{Validate build_prompt() Parameters}
\usage{
.validate_build_prompt(template)
}
\description{
Validate build_prompt() Parameters
}
\keyword{internal}
