% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_choice}
\alias{.validate_choice}
\title{Validate Choice from Set}
\usage{
.validate_choice(x, param_name, choices)
}
\arguments{
\item{x}{Value to check}

\item{param_name}{Name for error message}

\item{choices}{Valid choices}
}
\description{
Validate Choice from Set
}
\keyword{internal}
