% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_configure_tor_registry}
\alias{.validate_configure_tor_registry}
\title{Validate configure_tor_registry() Parameters}
\usage{
.validate_configure_tor_registry(
  registry_path,
  enable,
  bad_ttl,
  good_ttl,
  overuse_threshold,
  overuse_decay,
  max_rotation_attempts,
  ip_cache_ttl,
  conda_env
)
}
\description{
Validate configure_tor_registry() Parameters
}
\keyword{internal}
