% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_logical}
\alias{.validate_logical}
\title{Validate Boolean}
\usage{
.validate_logical(x, param_name)
}
\arguments{
\item{x}{Value to check}

\item{param_name}{Name for error message}
}
\description{
Validate Boolean
}
\keyword{internal}
