% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_output_format}
\alias{.validate_output_format}
\title{Validate output_format Argument}
\usage{
.validate_output_format(
  output_format,
  param_name = "output_format",
  fix = "Use output_format = \\"json\\""
)
}
\arguments{
\item{output_format}{Output format argument}

\item{param_name}{Name for error message}

\item{fix}{Hint text for invalid non-character values}
}
\description{
Validate output_format Argument
}
\keyword{internal}
