% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_process_outputs}
\alias{.validate_process_outputs}
\title{Validate process_outputs() Parameters}
\usage{
.validate_process_outputs(df, parallel, workers)
}
\description{
Validate process_outputs() Parameters
}
\keyword{internal}
