% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_s3_class}
\alias{.validate_s3_class}
\title{Validate S3 Class}
\usage{
.validate_s3_class(x, param_name, expected_class)
}
\arguments{
\item{x}{Value to check}

\item{param_name}{Name for error message}

\item{expected_class}{Expected S3 class name}
}
\description{
Validate S3 Class
}
\keyword{internal}
