% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_tor_options}
\alias{.validate_tor_options}
\title{Validate tor_options() Parameters}
\usage{
.validate_tor_options(tor, param_name = "tor")
}
\description{
Validate tor_options() Parameters
}
\keyword{internal}
