% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardrails__validate.R
\name{.validate_webpage_policy}
\alias{.validate_webpage_policy}
\title{Validate structured webpage policy argument}
\usage{
.validate_webpage_policy(webpage_policy, param_name = "webpage_policy")
}
\arguments{
\item{webpage_policy}{Value to validate}

\item{param_name}{Name for error messages}
}
\description{
Validate structured webpage policy argument
}
\keyword{internal}
