% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra__runtime_config.R
\name{.with_wayback_config}
\alias{.with_wayback_config}
\title{Temporarily Expose Wayback Runtime Settings to Python Tools}
\usage{
.with_wayback_config(wayback = NULL, fn)
}
\description{
Temporarily Expose Wayback Runtime Settings to Python Tools
}
\keyword{internal}
