% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-Coalescing Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left-hand side value}

\item{y}{Right-hand side value (default)}
}
\value{
x if not NULL, otherwise y
}
\description{
Returns the left-hand side if it is not NULL, otherwise returns the
right-hand side.
}
\keyword{internal}
