% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow__benchmarks.R
\name{list_benchmarks}
\alias{list_benchmarks}
\title{List Built-In ASA Benchmarks}
\usage{
list_benchmarks()
}
\value{
data.frame of benchmark metadata
}
\description{
Lists benchmark definitions bundled with the package.
}
