% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow__benchmarks.R
\name{load_benchmark}
\alias{load_benchmark}
\title{Load a Built-In ASA Benchmark}
\usage{
load_benchmark(name)
}
\arguments{
\item{name}{Benchmark name}
}
\value{
An object of class \code{asa_benchmark}
}
\description{
Loads a benchmark bundled under \code{inst/benchmarks/}.
}
