% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{truncate_string}
\alias{truncate_string}
\title{Truncate String}
\usage{
truncate_string(x, max_length = 100, ellipsis = "...")
}
\arguments{
\item{x}{Character string}

\item{max_length}{Maximum length}

\item{ellipsis}{String to append when truncated}
}
\value{
Truncated string
}
\description{
Truncates a string to a maximum length, adding ellipsis if truncated.
}
\keyword{internal}
