% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_ADAPTIVE_RATE_DECREASE}
\alias{ASA_ADAPTIVE_RATE_DECREASE}
\title{Adaptive Rate Decrease Factor (on success streak)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ASA_ADAPTIVE_RATE_DECREASE
}
\description{
Multiply delays by this factor after 10 consecutive successes.
}
\keyword{internal}
