% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_ADAPTIVE_RATE_ENABLED}
\alias{ASA_ADAPTIVE_RATE_ENABLED}
\title{Enable Adaptive Rate Limiting}
\format{
An object of class \code{logical} of length 1.
}
\usage{
ASA_ADAPTIVE_RATE_ENABLED
}
\description{
When TRUE, dynamically adjust delays based on success/error patterns.
}
\keyword{internal}
