% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_ADAPTIVE_RATE_INCREASE}
\alias{ASA_ADAPTIVE_RATE_INCREASE}
\title{Adaptive Rate Increase Factor (on error)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ASA_ADAPTIVE_RATE_INCREASE
}
\description{
Multiply delays by this factor when CAPTCHA/block detected.
}
\keyword{internal}
