% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_ADAPTIVE_RATE_WINDOW}
\alias{ASA_ADAPTIVE_RATE_WINDOW}
\title{Adaptive Rate Window Size (requests)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
ASA_ADAPTIVE_RATE_WINDOW
}
\description{
Number of recent requests to consider for adaptive rate adjustment.
}
\keyword{internal}
