% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_CIRCUIT_BREAKER_COOLDOWN}
\alias{ASA_CIRCUIT_BREAKER_COOLDOWN}
\title{Circuit Breaker Cooldown Period (seconds to wait when tripped)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
ASA_CIRCUIT_BREAKER_COOLDOWN
}
\description{
Circuit Breaker Cooldown Period (seconds to wait when tripped)
}
\keyword{internal}
