% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_CIRCUIT_BREAKER_ENABLED}
\alias{ASA_CIRCUIT_BREAKER_ENABLED}
\title{Enable Circuit Breaker by default}
\format{
An object of class \code{logical} of length 1.
}
\usage{
ASA_CIRCUIT_BREAKER_ENABLED
}
\description{
Enable Circuit Breaker by default
}
\keyword{internal}
