% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_CIRCUIT_BREAKER_THRESHOLD}
\alias{ASA_CIRCUIT_BREAKER_THRESHOLD}
\title{Circuit Breaker Error Threshold (trip if error rate exceeds this)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ASA_CIRCUIT_BREAKER_THRESHOLD
}
\description{
Circuit Breaker Error Threshold (trip if error rate exceeds this)
}
\keyword{internal}
