% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_CIRCUIT_BREAKER_WINDOW}
\alias{ASA_CIRCUIT_BREAKER_WINDOW}
\title{Circuit Breaker Window Size (number of recent requests to consider)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
ASA_CIRCUIT_BREAKER_WINDOW
}
\description{
Circuit Breaker Window Size (number of recent requests to consider)
}
\keyword{internal}
