% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\name{ASA_DEFAULT_AZURE_OPENAI_MODEL}
\alias{ASA_DEFAULT_AZURE_OPENAI_MODEL}
\title{Default Azure OpenAI Deployment}
\usage{
ASA_DEFAULT_AZURE_OPENAI_MODEL
}
\description{
Default Azure OpenAI Deployment
}
\keyword{internal}
