% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_DEFAULT_INTER_SEARCH_DELAY}
\alias{ASA_DEFAULT_INTER_SEARCH_DELAY}
\title{Default Inter-Search Delay (seconds)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ASA_DEFAULT_INTER_SEARCH_DELAY
}
\description{
Default Inter-Search Delay (seconds)
}
\keyword{internal}
