% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_DEFAULT_INVOKE_MAX_ATTEMPTS}
\alias{ASA_DEFAULT_INVOKE_MAX_ATTEMPTS}
\title{Default Agent Invoke Retry Attempts}
\format{
An object of class \code{integer} of length 1.
}
\usage{
ASA_DEFAULT_INVOKE_MAX_ATTEMPTS
}
\description{
Total attempts for transient model invoke failures.
}
\keyword{internal}
