% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_DEFAULT_INVOKE_RETRY_DELAY}
\alias{ASA_DEFAULT_INVOKE_RETRY_DELAY}
\title{Default Agent Invoke Retry Delay (seconds)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ASA_DEFAULT_INVOKE_RETRY_DELAY
}
\description{
Default Agent Invoke Retry Delay (seconds)
}
\keyword{internal}
