% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\docType{data}
\name{ASA_DEFAULT_MEMORY_SUMMARIZER_MAX_OUTPUT_TOKENS}
\alias{ASA_DEFAULT_MEMORY_SUMMARIZER_MAX_OUTPUT_TOKENS}
\title{Default Memory Summarizer Output Cap (tokens)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
ASA_DEFAULT_MEMORY_SUMMARIZER_MAX_OUTPUT_TOKENS
}
\description{
Default Memory Summarizer Output Cap (tokens)
}
\keyword{internal}
