% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_DEFAULT_MEMORY_THRESHOLD}
\alias{ASA_DEFAULT_MEMORY_THRESHOLD}
\title{Default Memory Threshold (messages before folding)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
ASA_DEFAULT_MEMORY_THRESHOLD
}
\description{
Default Memory Threshold (messages before folding)
}
\keyword{internal}
