% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_DEFAULT_OM_BLOCK_AFTER}
\alias{ASA_DEFAULT_OM_BLOCK_AFTER}
\title{Default OM Block-After Ratio}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ASA_DEFAULT_OM_BLOCK_AFTER
}
\description{
Default OM Block-After Ratio
}
\keyword{internal}
