% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_DEFAULT_OM_BUFFER_ACTIVATION}
\alias{ASA_DEFAULT_OM_BUFFER_ACTIVATION}
\title{Default OM Buffer Activation Ratio}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ASA_DEFAULT_OM_BUFFER_ACTIVATION
}
\description{
Default OM Buffer Activation Ratio
}
\keyword{internal}
