% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\docType{data}
\name{ASA_DEFAULT_OM_BUFFER_TOKENS}
\alias{ASA_DEFAULT_OM_BUFFER_TOKENS}
\title{Default OM Buffered Observation Budget (approx tokens)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
ASA_DEFAULT_OM_BUFFER_TOKENS
}
\description{
Default OM Buffered Observation Budget (approx tokens)
}
\keyword{internal}
