% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_DEFAULT_PLATEAU_ROUNDS}
\alias{ASA_DEFAULT_PLATEAU_ROUNDS}
\title{Default Plateau Rounds for Stopping}
\format{
An object of class \code{integer} of length 1.
}
\usage{
ASA_DEFAULT_PLATEAU_ROUNDS
}
\description{
Default Plateau Rounds for Stopping
}
\keyword{internal}
