% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_DEFAULT_RATE_LIMIT}
\alias{ASA_DEFAULT_RATE_LIMIT}
\title{Default Rate Limit (requests per second)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ASA_DEFAULT_RATE_LIMIT
}
\description{
Conservative default: 0.1 = 10 seconds between requests.
  Tuned for heavy volume (1000+ queries/day) to reduce CAPTCHA/blocks.
}
\keyword{internal}
