% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_DEFAULT_SEARCH_TIMEOUT}
\alias{ASA_DEFAULT_SEARCH_TIMEOUT}
\title{Default Search Timeout (seconds)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ASA_DEFAULT_SEARCH_TIMEOUT
}
\description{
Per-search HTTP timeout used by search/network tools.
}
\keyword{internal}
