% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\docType{data}
\name{ASA_DEFAULT_SELENIUM_BROWSER_PREFERENCE}
\alias{ASA_DEFAULT_SELENIUM_BROWSER_PREFERENCE}
\title{Default Selenium Browser Preference}
\format{
An object of class \code{character} of length 1.
}
\usage{
ASA_DEFAULT_SELENIUM_BROWSER_PREFERENCE
}
\description{
Controls Selenium engine order in browser tier.
  "firefox_first" uses Firefox before Chrome fallbacks.
}
\keyword{internal}
