% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_DEFAULT_USE_BROWSER}
\alias{ASA_DEFAULT_USE_BROWSER}
\title{Default Use Browser (Selenium Tier Enabled)}
\format{
An object of class \code{logical} of length 1.
}
\usage{
ASA_DEFAULT_USE_BROWSER
}
\description{
Default Use Browser (Selenium Tier Enabled)
}
\keyword{internal}
