% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\docType{data}
\name{ASA_DEFAULT_USE_OBSERVATIONAL_MEMORY}
\alias{ASA_DEFAULT_USE_OBSERVATIONAL_MEMORY}
\title{Default Observational Memory Enabled}
\format{
An object of class \code{logical} of length 1.
}
\usage{
ASA_DEFAULT_USE_OBSERVATIONAL_MEMORY
}
\description{
Default Observational Memory Enabled
}
\keyword{internal}
