% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_HUMANIZE_TIMING}
\alias{ASA_HUMANIZE_TIMING}
\title{Enable Humanized Timing (random jitter on delays)}
\format{
An object of class \code{logical} of length 1.
}
\usage{
ASA_HUMANIZE_TIMING
}
\description{
Enable Humanized Timing (random jitter on delays)
}
\keyword{internal}
