% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_JITTER_FACTOR}
\alias{ASA_JITTER_FACTOR}
\title{Jitter Factor for random timing variation}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ASA_JITTER_FACTOR
}
\description{
Jitter Factor for random timing variation
}
\keyword{internal}
