% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\docType{data}
\name{ASA_OPENCODE_DEFAULT_AGENT}
\alias{ASA_OPENCODE_DEFAULT_AGENT}
\title{Default OpenCode Agent Name}
\format{
An object of class \code{character} of length 1.
}
\usage{
ASA_OPENCODE_DEFAULT_AGENT
}
\description{
Default OpenCode Agent Name
}
\keyword{internal}
