% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_PROACTIVE_ROTATION_INTERVAL}
\alias{ASA_PROACTIVE_ROTATION_INTERVAL}
\title{Proactive Rotation Interval (requests)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
ASA_PROACTIVE_ROTATION_INTERVAL
}
\description{
Rotate Tor circuit every 15 requests to get fresh exit node IP.
}
\keyword{internal}
