% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_RATE_LIMIT_BUCKET_SIZE}
\alias{ASA_RATE_LIMIT_BUCKET_SIZE}
\title{Proactive Rate Limit Bucket Size (max tokens)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
ASA_RATE_LIMIT_BUCKET_SIZE
}
\description{
Proactive Rate Limit Bucket Size (max tokens)
}
\keyword{internal}
