% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_RATE_LIMIT_PROACTIVE}
\alias{ASA_RATE_LIMIT_PROACTIVE}
\title{Enable Proactive Rate Limiting (default: TRUE)}
\format{
An object of class \code{logical} of length 1.
}
\usage{
ASA_RATE_LIMIT_PROACTIVE
}
\description{
Enable Proactive Rate Limiting (default: TRUE)
}
\keyword{internal}
