% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_RECURSION_LIMIT_FOLDING}
\alias{ASA_RECURSION_LIMIT_FOLDING}
\title{Recursion Limit with Memory Folding}
\format{
An object of class \code{integer} of length 1.
}
\usage{
ASA_RECURSION_LIMIT_FOLDING
}
\description{
Recursion Limit with Memory Folding
}
\keyword{internal}
