% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_SESSION_RESET_INTERVAL}
\alias{ASA_SESSION_RESET_INTERVAL}
\title{Session Reset Interval (requests)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
ASA_SESSION_RESET_INTERVAL
}
\description{
Reset session identity every 50 requests (clear cookies, shuffle UA).
}
\keyword{internal}
