% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__constants.R
\docType{data}
\name{ASA_SUPPORTED_AGENT_BACKENDS}
\alias{ASA_SUPPORTED_AGENT_BACKENDS}
\title{Supported Agent Backends}
\format{
An object of class \code{character} of length 2.
}
\usage{
ASA_SUPPORTED_AGENT_BACKENDS
}
\description{
Supported Agent Backends
}
\keyword{internal}
