% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_TEMPORAL_STRICTNESS}
\alias{ASA_TEMPORAL_STRICTNESS}
\title{Valid Temporal Strictness Levels}
\format{
An object of class \code{character} of length 2.
}
\usage{
ASA_TEMPORAL_STRICTNESS
}
\description{
Valid Temporal Strictness Levels
}
\keyword{internal}
