% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_TOR_GOOD_TTL}
\alias{ASA_TOR_GOOD_TTL}
\title{Tor Exit Good TTL (seconds)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ASA_TOR_GOOD_TTL
}
\description{
How long to treat an exit as good before requiring a refresh.
}
\keyword{internal}
