% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_TOR_MIN_ROTATION_INTERVAL}
\alias{ASA_TOR_MIN_ROTATION_INTERVAL}
\title{Minimum Tor Rotation Interval (seconds)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ASA_TOR_MIN_ROTATION_INTERVAL
}
\description{
Minimum time between Tor circuit rotations to avoid hammering.
}
\keyword{internal}
