% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_TOR_OVERUSE_DECAY}
\alias{ASA_TOR_OVERUSE_DECAY}
\title{Tor Exit Overuse Decay Window (seconds)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ASA_TOR_OVERUSE_DECAY
}
\description{
Time window for counting recent uses before decaying counts.
}
\keyword{internal}
