% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_TOR_OVERUSE_THRESHOLD}
\alias{ASA_TOR_OVERUSE_THRESHOLD}
\title{Tor Exit Overuse Threshold}
\format{
An object of class \code{integer} of length 1.
}
\usage{
ASA_TOR_OVERUSE_THRESHOLD
}
\description{
Maximum recent uses before a good exit is considered overloaded.
}
\keyword{internal}
