% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ASA_TOR_REGISTRY_ENABLED}
\alias{ASA_TOR_REGISTRY_ENABLED}
\title{Enable Shared Tor Exit Registry}
\format{
An object of class \code{logical} of length 1.
}
\usage{
ASA_TOR_REGISTRY_ENABLED
}
\description{
When TRUE, track Tor exit IP health (good/bad/overused) in a shared store.
}
\keyword{internal}
