% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_agent.R
\name{.create_http_clients}
\alias{.create_http_clients}
\title{Create HTTP Clients for API Calls}
\usage{
.create_http_clients(proxy, timeout)
}
\arguments{
\item{proxy}{Proxy URL or NULL}

\item{timeout}{Timeout in seconds}
}
\description{
Create HTTP Clients for API Calls
}
\keyword{internal}
