% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_agent.R
\name{.create_tools}
\alias{.create_tools}
\title{Create Search Tools}
\usage{
.create_tools(proxy)
}
\arguments{
\item{proxy}{Proxy URL or NULL}
}
\description{
Create Search Tools
}
\keyword{internal}
