% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_agent.R
\name{.invoke_standard_agent}
\alias{.invoke_standard_agent}
\title{Invoke Standard Agent}
\usage{
.invoke_standard_agent(python_agent, prompt, recursion_limit, thread_id = NULL)
}
\description{
Invoke Standard Agent
}
\keyword{internal}
