% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core__package_state.R
\name{.is_initialized}
\alias{.is_initialized}
\title{Check if ASA Agent is Initialized}
\usage{
.is_initialized()
}
\value{
Logical indicating if the agent has been initialized
}
\description{
Check if ASA Agent is Initialized
}
\keyword{internal}
