% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_task.R
\name{.parse_json_response}
\alias{.parse_json_response}
\title{Parse JSON Response}
\usage{
.parse_json_response(response_text)
}
\arguments{
\item{response_text}{Response text from agent}
}
\description{
Parse JSON Response
}
\keyword{internal}
