% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow__research_enumeration.R
\name{.save_checkpoint}
\alias{.save_checkpoint}
\title{Save Checkpoint}
\usage{
.save_checkpoint(result, query, schema_dict, config_dict, checkpoint_file)
}
\description{
Save Checkpoint
}
\keyword{internal}
