% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_tor_ip}
\alias{get_tor_ip}
\title{Get External IP via Tor}
\usage{
get_tor_ip(proxy = "socks5h://127.0.0.1:9050")
}
\arguments{
\item{proxy}{Tor proxy URL}
}
\value{
IP address string or NA on failure
}
\description{
Retrieves the external IP address as seen through Tor proxy.
}
\examples{
\dontrun{
ip <- get_tor_ip()
message("Current Tor IP: ", ip)
}

}
